+++ onnx_safetensors/_safetensors_io.py
